- refactor to add NucleosideSequenceItemRenderer class
