+++ packages/ketcher-core/src/application/render/renderers/sequence/NucleotideSequenceItemRenderer.ts
@@ -1,48 +1,21 @@
-import { BaseSequenceItemRenderer } from 'application/render/renderers/sequence/BaseSequenceItemRenderer';
 import { Nucleotide } from 'domain/entities';
 import { D3SvgElementSelection } from 'application/render/types';
 
-export class NucleotideSequenceItemRenderer extends BaseSequenceItemRenderer {
+import { RNASequenceItemRenderer } from './RNASequenceItemRenderer';
+
+export class NucleotideSequenceItemRenderer extends RNASequenceItemRenderer {
   private phosphateModificationCircleElement?: D3SvgElementSelection<
     SVGCircleElement,
     void
   >;
 
-  get symbolToDisplay(): string {
-    return (
-      this.node.monomer.attachmentPointsToBonds.R3?.getAnotherMonomer(
-        this.node.monomer,
-      )?.monomerItem?.props.MonomerNaturalAnalogCode || '@'
-    );
-  }
-
-  protected drawModification() {
+  drawModification() {
     const node = this.node as Nucleotide;
+
+    this.drawCommonModification(node);
 
     if (this.phosphateModificationCircleElement) {
       this.phosphateModificationCircleElement.remove();
-    }
-
-    if (node.rnaBase.isModification) {
-      this.backgroundElement?.attr(
-        'fill',
-        this.node.monomer.selected
-          ? this.isSequenceEditInRnaBuilderModeTurnedOn
-            ? '#41A8B2'
-            : '#3ACA6A'
-          : this.isSequenceEditModeTurnedOn
-          ? '#ff7a004f'
-          : '#CAD3DD',
-      );
-    }
-
-    if (node.sugar.isModification) {
-      this.backgroundElement
-        ?.attr(
-          'stroke',
-          this.isSequenceEditInRnaBuilderModeTurnedOn ? '#24545A' : '#585858',
-        )
-        .attr('stroke-width', '1px');
     }
 
     if (node.phosphate?.isModification) {
@@ -53,21 +26,6 @@ export class NucleotideSequenceItemRenderer extends BaseSequenceItemRenderer {
           'fill',
           this.isSequenceEditInRnaBuilderModeTurnedOn ? '#24545A' : '#585858',
         )
-        .attr('cx', '10')
-        .attr('cy', '-16');
-    }
-
-    // show modification for not last Nucleoside
-    if (!node.phosphate) {
-      this.phosphateModificationCircleElement = this.rootElement
-        ?.append('circle')
-        .attr('r', '3px')
-        .attr(
-          'stroke',
-          this.isSequenceEditInRnaBuilderModeTurnedOn ? '#24545A' : '#585858',
-        )
-        .attr('stroke-width', '1px')
-        .attr('fill', 'none')
         .attr('cx', '10')
         .attr('cy', '-16');
     }
